Add warnings for ambiguous / unknown reviewers
Added warning messages when some of the reviewers in an uploaded
batch group list are ambiguous e.g. there are multiple matches for
their keys, or when reviewers with given keys are not found.

Also removed a check that skipped reviewers with lti_user_id - if
there are LTI users that also have non-lti accounts with the same
identifiers, they're skipped in assigning reviewers anyway. The
old functionality prevented LTI users from being assigned even
if their identifiers were unique.